Pass param_handler_manager to runner factory, and simplify interface with get/instantiate methods. Move Runner and RunnerBackend to avoid circular imports.
