+++ labtech/runners/__init__.py
@@ -1,8 +1,11 @@
+from .base import Runner, RunnerBackend
 from .process import ForkRunnerBackend, SpawnRunnerBackend
 from .serial import SerialRunnerBackend
 from .thread import ThreadRunnerBackend
 
 __all__ = [
+    'Runner',
+    'RunnerBackend',
     'ForkRunnerBackend',
     'SpawnRunnerBackend',
     'SerialRunnerBackend',
