+++ labtech/types.py
@@ -9,7 +9,6 @@ from typing import (
     Any,
     Callable,
     Generic,
-    Iterator,
     Literal,
     Optional,
     Protocol,
@@ -275,7 +274,7 @@ class ParamHandler(Protocol):
         """Returns True if the given parameter value should be handled
         by this class."""
 
-    def find_tasks(self, value: Any, *, find_tasks_in_param: Callable[[Any], Sequence[Task]]) -> Sequence[Task]:
+    def find_tasks(self, value: Any, *, find_tasks_in_param: Callable[[Any], Sequence[Task]]) -> list[Task]:
         """Given a parameter value, return all tasks within it (not
         including tasks within those tasks).
 
@@ -303,123 +302,3 @@ class ParamHandler(Protocol):
 TaskMonitorInfoValue = datetime | str | int | float
 TaskMonitorInfoItem = TaskMonitorInfoValue | tuple[TaskMonitorInfoValue, str]
 TaskMonitorInfo = dict[str, TaskMonitorInfoItem]
-
-
-class Runner(ABC):
-    """Manages the execution of [Tasks][labtech.types.Task], typically
-    by delegating to a parallel processing framework."""
-
-    @abstractmethod
-    def submit_task(self, task: Task, task_name: str, use_cache: bool) -> None:
-        """Submit the given task object to be run and have its result cached.
-
-        It is up to the Runner to decide when to start running the
-        task (i.e. when resources become available).
-
-        The implementation of this method should run the task by
-        effectively calling:
-
-        ```
-        # param_handler_manager should be saved from
-        # labtech.params.get_param_handler_manager() on the main process
-        # and set in remote processes that don't inherit from the main process:
-        labtech.params.set_param_handler_manager(param_handler_manager)
-
-        for dependency_task in get_direct_dependencies(task, all_identities=True):
-            # Where results_map is expected to contain the TaskResult for
-            # each dependency_task.
-            dependency_task._set_results_map(results_map)
-
-        current_process = multiprocessing.current_process()
-        orig_process_name = current_process.name
-        try:
-            # If the thread name or similar is set instead of the process
-            # name, then the Runner should update the handler of the global
-            # labtech.utils.logger to include that instead of the process name.
-            current_process.name = task_name
-            return labtech.runners.base.run_or_load_task(
-                task=task,
-                use_cache=use_cache,
-                filtered_context=task.filter_context(context),
-                storage=storage,
-            )
-        finally:
-            current_process.name = orig_process_name
-        ```
-
-        Args:
-            task: The task to execute.
-            task_name: Name to use when referring to the task in logs.
-            use_cache: If True, the task's result should be fetched from the
-                cache if it is available (fetching should still be done in a
-                delegated process).
-
-        """
-
-    @abstractmethod
-    def wait(self, *, timeout_seconds: Optional[float]) -> Iterator[tuple[Task, ResultMeta | BaseException]]:
-        """Wait up to timeout_seconds or until at least one of the
-        submitted tasks is done, then return an iterator of tasks in a
-        done state and a list of tasks in all other states.
-
-        Each task is returned as a pair where the first value is the
-        task itself, and the second value is either:
-
-        * For a successfully completed task: Metadata of the result.
-        * For a task that fails with any BaseException descendant: The exception
-          that was raised.
-
-        Cancelled tasks are never returned.
-
-        """
-
-    @abstractmethod
-    def cancel(self) -> None:
-        """Cancel all submitted tasks that have not yet been started."""
-
-    @abstractmethod
-    def stop(self) -> None:
-        """Stop all currently running tasks."""
-
-    @abstractmethod
-    def close(self) -> None:
-        """Clean up any resources used by the Runner after all tasks
-        are finished, cancelled, or stopped."""
-
-    @abstractmethod
-    def pending_task_count(self) -> int:
-        """Returns the number of tasks that have been submitted but
-        not yet cancelled or returned from a call to wait()."""
-
-    @abstractmethod
-    def get_result(self, task: Task) -> TaskResult:
-        """Returns the in-memory result for a task that was
-        successfully run by this Runner. Raises a KeyError for a
-        result with no in-memory result."""
-
-    @abstractmethod
-    def remove_results(self, tasks: Sequence[Task]) -> None:
-        """Removes the in-memory results for tasks that were
-        sucessfully run by this Runner. Ignores tasks that have no
-        in-memory result."""
-
-    @abstractmethod
-    def get_task_infos(self) -> list[TaskMonitorInfo]:
-        """Returns a snapshot of monitoring information about each
-        task that is currently running."""
-
-
-class RunnerBackend(ABC):
-    """Factory class to construct [Runner][labtech.types.Runner] objects."""
-
-    @abstractmethod
-    def build_runner(self, *, context: LabContext, storage: Storage, max_workers: Optional[int]) -> Runner:
-        """Return a Runner prepared with the given configuration.
-
-        Args:
-            context: Additional variables made available to tasks that aren't
-                considered when saving to/loading from the cache.
-            storage: Where task results should be cached to.
-            max_workers: The maximum number of parallel worker processes for
-                running tasks.
-        """
